basic-agent/pyproject.toml
[project]
name = "basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
